Move `compile_catalog_plusjs` to `utils/`
